Fitured(Item) Add CRUD after Validation

+++ Controllers/ItemController.cs
@@ -2,8 +2,6 @@
 using TrainingProjectAPI.Models.DB;
 using TrainingProjectAPI.Services;
 
-// For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
-
 namespace TrainingProjectAPI.Controllers
 {
     [Route("api/[controller]")]
@@ -17,40 +15,90 @@ namespace TrainingProjectAPI.Controllers
             _itemService = itemService;
         }
 
-
-        // GET: api/<CustomerController>
+        // GET: api/<ItemController>
         [HttpGet]
-        public IActionResult Get()
+        public ActionResult GetItem()
         {
             var ItemList = _itemService.GetListItem();
             return Ok(ItemList);
         }
 
-        // POST api/<CustomerController>
+        // POST api/<ItemController>
         [HttpPost]
-        public IActionResult Post(Item item)
+        public IActionResult CreateItem([FromBody] Item item)
         {
-            var addItem = _itemService.CreateItem(item);
-            if (addItem)
+            if (string.IsNullOrWhiteSpace(item.NamaItem))
+                return BadRequest("Nama Item tidak boleh kosong atau hanya spasi");
+
+            if (item.NamaItem.Length > 100)
+                return BadRequest("Nama Item Maksimal 100 karakter");
+
+            if (item.QTY <= 0)
+                return BadRequest("QTY harus lebih dari 0");
+            
+            if (item.TglExpired <= DateTime.Now)
+                return BadRequest("Tanggal harus lebih besar dari hari ini");
+
+            if (string.IsNullOrWhiteSpace(item.Supplier))
+                return BadRequest("Supplier tidak boleh kosong atau hanya spasi");
+
+            if (item.Supplier.Length > 100)
+                return BadRequest("Supplier Maksimal 100 karakter");
+
+            if (item.AlamatSupplier.Length > 100)
+                return BadRequest("Alamat Supplier Maksimal 100 Karakter!");
+
+            var cekNamaItem = _itemService.GetListItem().FirstOrDefault(x => x.NamaItem == item.NamaItem);
+            if (cekNamaItem != null)
+                return BadRequest("Nama Item sudah ada");
+
+            var insertItem = _itemService.CreateItem(item);
+            if (insertItem)
             {
-                return Ok("Insert Item Succes");
+                return Ok("Insert Item Success");
             }
-            return BadRequest("insert Item Failed");
+            return BadRequest("Insert Item Failed");
         }
 
-        // PUT api/<CustomerController>/5
+        // PUT api/<ItemController>
         [HttpPut]
-        public IActionResult Put(Item item)
+        public IActionResult UpdateItem([FromBody] Item item)
         {
+            if (string.IsNullOrWhiteSpace(item.NamaItem))
+                return BadRequest("Nama Item tidak boleh kosong atau hanya spasi");
+
+            if (item.NamaItem.Length > 100)
+                return BadRequest("Nama Item tidak boleh lebih dari 100 karakter");
+
+            if (item.QTY < 0)
+                return BadRequest("QTY harus lebih dari 0");
+
+            if (item.TglExpired <= DateTime.Now)
+                return BadRequest("Tanggal kadaluarsa harus lebih besar dari hari ini");
+
+            if (string.IsNullOrWhiteSpace(item.Supplier))
+                return BadRequest("Supplier tidak boleh kosong atau hanya spasi");
+
+            if (item.Supplier.Length > 100)
+                return BadRequest("Supplier Maksimal 100 karakter");
+
+            var cekNamaItem = _itemService.GetListItem().FirstOrDefault(x => x.NamaItem == item.NamaItem && x.Id != item.Id);
+            if (cekNamaItem != null)
+                return BadRequest("Nama Item sudah ada");
+
+            var existingItem = _itemService.GetListItem().FirstOrDefault(existId => existId.Id == item.Id);
+            if (existingItem == null)
+                return NotFound("Item dengan ID tersebut tidak ditemukan");
+
             try
             {
                 var updateItem = _itemService.UpdateItem(item);
                 if (updateItem)
                 {
-                    return Ok("Update Item Succes");
+                    return Ok("Update Item Success");
                 }
 
-                return BadRequest("insert Item Failed");
+                return BadRequest("Update Item Failed");
             }
             catch (Exception ex)
             {
@@ -59,10 +107,13 @@ namespace TrainingProjectAPI.Controllers
             }
         }
 
-        // DELETE api/<CustomerController>/5
+        // DELETE api/<ItemController>/5
         [HttpDelete]
-        public IActionResult Delete(int id)
+        public IActionResult DeleteItem(int id)
         {
+            if (id <= 0)
+                return BadRequest("ID tidak valid");
+
             try
             {
                 var deleteItem = _itemService.DeleteItem(id);
